backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Data Sources
yfinance>=0.2.51
fredapi>=0.5.2
finnhub-python>=2.4.20
alpha-vantage>=3.0.0

# Analysis
pandas>=2.0.0
ta>=0.11.0
numpy>=1.26.4
scipy>=1.15.1
hmmlearn>=0.3.3
scikit-learn>=1.6.1

# AI
anthropic>=0.43.0

# Scheduling
apscheduler>=3.10.4

# Utilities
httpx>=0.28.1
python-dotenv>=1.0.1
python-multipart>=0.0.22
